Refactor mouse interaction into separate MouseInteractionManager module
